chore: move files & delete unnecessary css & edit name
