+++ src/app/[year]/moa/mymoa/(components)/(ui)/GuideModalMiddle.tsx
