+++ src/styles/create-moa/createMoa.module.css
@@ -178,16 +178,6 @@
   }
 }
 
-.step3_container {
-  margin-left: 40px;
-  margin-right: 50px;
-  display: flex;
-  flex-direction: column;
-  max-width: 100vw;
-  height: 75vh;
-  position: relative;
-}
-
 .width {
   width: 100%;
   margin-right: 50px;
